# apis4lawenforcement_swaggerdef.yml
swagger: '2.0'
info:
  version: '1.0'
  title: "APIs 4 SAPOL Demo"
  description: SAPol provides a simple API allowing consumers to view people, license, vehicle and property info.
  #license:
  #  name: MIT
  #  url: https://github.com/apiaryio/polls-api/blob/master/LICENSE
...
basePath: "/"
host: "localhost:3000"
...
schemes:
- http
security:
  - X-MyAPI-Key: []
consumes:
- application/json
produces:
- application/json
paths:
...
  /officers:
    x-summary: Officer Lookup 
    get:
      tags: 
      - "Officers"
      summary: Police Officer Details
      description: "Returns information for a single Police Officer"
      operationId: "getDetailsForOfficer"
      produces:
      - "application/json"
      x-summary: Officer Details
      parameters:
      - name: "fname"
        in: "query"
        description: "firstName of the contact used to filter the record(s) to be returned."
        required: false
        type: "string"
        x-example: 'Dave'
      - name: "lname"
        in: "query"
        description: "lastName of the contact used to filter the record(s) to be returned."
        required: false
        type: "string"
        x-example: 'Smith'
      - name: "dob"
        in: "query"
        description: "date of birth of the contact used to filter the record(s) to be returned."
        required: false
        type: "string"
        x-example: '12-DEC-1980'
      responses:
        200:
          description: Successful Task Lookup Response
          schema:
            properties:
              Officers:
                type: array
                items:
                  $ref: '#/definitions/PoliceOfficerResponse'
          examples:
            application/json:
              Officers:
                fname: 'Dave'
                lname: 'Redbull'
                dob: '12-JUN-1980'
                address: '20 Charlesworthnut Street, Walkerville, SA, 5000'
                addresslat: -34.778011
                addresslong: 138.667145
                badgeId: 'A12345'
                rank: 'Police Constable'
                assignedToLSA: 'Adelaide'
  /officers/{badgeId}:
    put:
      tags:
      - "Officers"
      summary: 'Update the details of an Officer'
      description: 'Update Officer details eg assign to new LSA.'
      operationId: "updateOfficerByBadgeId"
      produces:
      - "application/json"
      parameters:
        - name: 'badgeId'
          in: path
          required: true
          type: string
          x-example: 'B99999'
        - name: body
          in: body
          required: true
          schema:
            properties:
              PoliceOfficer:
                $ref: '#/definitions/PoliceOfficerRequest'
      responses:
        201:
          description: 'Successful update of Officer Information'
          schema:
            $ref: '#/definitions/PoliceOfficerResponse'
          examples:
            application/json:
              PoliceOfficer:
                id: 1
                fname: 'Dave'
                lname: 'Redbull'
                dob: '01-JAN-1980'
                address: '20 Charlesworthnut Street, Walkerville, SA 5000'
                addresslat: -34.778011
                addresslong: 138.667145
                badgeId: 'A12345'
                rank: 'Police Constable'
                assignedToLSA: 'Adelaide'
...
securityDefinitions:
  X-MyAPI-Key:
    type: "apiKey"
    name: "x-api-key"
    in: "header"
definitions:
...
  PoliceOfficerRequest:
    title: "Police Officer Request"
    required:
    - "fname"
    - "lname"
    - "badgeId"
    type: object
    properties:
      fname:
        type: string
        example: 'Michael'
      lname:
        type: string
        example: 'Mouse'
      dob:
        type: string
        example: '01-JAN-2001'
      address:
        type: string
        example: 'No Fixed Address'
      addresslat:
        type: number
        example: -34.959247
      addresslong:
        type: number
        example: 138.638071
      badgeId:
        type: string
        example: 'B99988'
      rank:
        type: string
        example: 'Police Constable'
      assignedtoLSA:
        type: string
        example: 'Adelaide'
  PoliceOfficerResponse:
    title: "Police Officer Response"
    type: object
    properties:
      fname:
        type: string
        example: 'Michael'
      lname:
        type: string
        example: 'Mouse'
      dob:
        type: string
        example: '01-JAN-2001'
      address:
        type: string
        example: 'No Fixed Address'
      addresslat:
        type: number
        example: -34.959247
      addresslong:
        type: number
        example: 138.638071
      badgeId:
        type: string
        example: 'B99988'
      rank:
        type: string
        example: 'Police Constable'
      assignedtoLSA:
        type: string
        example: 'Adelaide'
...